template/requirements.txt
croniter
marshmallow
marshmallow-dataclass
marshmallow-enum
python-dateutil
python-telegram-bot
pytz
PyYAML
